ints to floats

+++ Assignment2/src/StatisticsCalculator.java
@@ -50,9 +50,9 @@ public class StatisticsCalculator {
      * @return
      */
     public static String RunStatistics(ArrayList<Process> list){
-	int averageTurnAroundTime = 0;
-	int averageResponceTime = 0;
-	int averageWaitTime = 0;
+	float averageTurnAroundTime = 0;
+	float averageResponceTime = 0;
+	float averageWaitTime = 0;
 	for (Process process: list){
 	    averageWaitTime += process.getResponseTime();
 	    averageTurnAroundTime += process.getTurnAroundTime();
